requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
gunicorn>=21.0.0
whitenoise>=6.5.0
python-decouple>=3.8
Pillow>=10.0.0